Add hue-shift control for the channel mixer.
This shit the hue for the red, green or bleu individually. Those new
controls are designed to be used for color-grading work.

When a input R/G/B slider is moved the new hue-shift slider is marked
as undeding as there is no way back from R/G/B to hue-shift (most
values have no correspondances).