Seems to work better.
When a distort module is used we need to invalidate the cached
tone-equalizer luminance mask. To do that we introduce a new signal
DT_SIGNAL_DEVELOP_DISTORT which is raised when a module doing distortion
is enabled/disabled or its parameter changed. The tone-equalizer module
ses this signal to invalidate the internal cache and to get a new one
corresponding to the actual development.

This is a long standing issue and can be reproduced with many modules (flip,
orientation, rotation & perspective). One way to reproduce:

1. Open a landscape picture
2. Enable tone-equalizer, fly over and the that the luminance indicator
   corresponds to the actual on-canvas value.
3. Enable flip (no cache, image will be recomputed, new cached mask taken
   by tone-equalizer)
4. Disable flip  (the previous image is in the cache, pipe is not run)
5. Fly over the image and see the luminance not being correctly reported.

This is due to the fact that the last cached mask in from 3 but the actual
on-canvas view does not corresponds.

It is harder to reproduce with the lens correction as the deformation is
light. But it is possible to reproduce with rotation & perspective with
a 10% rotation for example.